Change humidity to uint16_t

+++ include/payload.h
@@ -30,7 +30,7 @@ typedef struct
 } action_t;
 
 payload_id_t     payload_get_id(const char *hex_str);
-payload_uplink_t payload_pack_thc(uint8_t flags, int16_t temperature, uint8_t humidity, uint16_t co2);
+payload_uplink_t payload_pack_thc(uint8_t flags, int16_t temperature, uint16_t humidity, uint16_t co2);
 
 void payload_unpack_thc_presets(const char *hex_str, range_t *temp_range, range_t *hum_range, range_t *co2_range);
 void payload_unpack_actions(const char *hex_str, action_t *actions);
